Moved the whole addEffects method to the FoodDataBase
There's a new class called EffectValue which contains all the
information about the effects a food item contains. Those effectvalues
are saved in the FoodValue-class, which is saved in the FoodDataBase.
This makes the getting and setting of effects way easier and the config
file doesn't have to be called all the time.