Checking, if sender is player
In the "get" command, if one doesn't specify the item he's looking for,
it looks in the inventory of the player. Previously the sender was just
casted to a player, but this could lead to a little problem if the
console is not specifying the item. And the console also doesn't have a
freaking inventory!