Instead of storing the potion effect type, it stores an int
Originally the EffectValue stored the PotionEffectType as effect. I
changed it to an int, because I wanted to include giving XP levels that
way. Since there are no PotionEffects that give XP values, I'm just
cheating a little bit by saving the int. 0 means that the effect is
giving XP values, everything else stays the same.